refactor: extract tournament logic
